feat(pages and components): implement routing functionality between pages and increase item list for reports page
